// src/app/students/model/addstudent.ts
export interface AddStudentData
{
  academicYear:  string;
    firstName:   string;
    middleName:  string;
    lastName :   string;
    className:   string;
    sectionName: string;
    rollNo:      string;
    gender:      string;
    email:       string;
    mobile:      string;
  admissionDate: string;
  landlineNumber:string;
    parentName:  string;
    parentNumber:string;
    studentDOB:  string;
    bGroup:     string;
    address:    string;
    uploadImg:   string;
}